Se agregaron todoslos elementos del mapa incidencias
