+++ incidencias/src/app/app.component.html
@@ -1,19 +1,5 @@
-<!-- <div class="d-flex" id="wrapper">
-    <div class="bg-light border-right " id="sidebar-wrapper">
-        <div class="sidebar-heading">Incidencias de la CDMX</div>
-        <div class="list-group list-group-flush">
-            <a class="list-group-item list-group-item-action bg-light" [routerLink]="['map']"> Mapa </a>
-            <a class="list-group-item list-group-item-action bg-light" [routerLink]="['chart']">Estadisticas </a>
-        </div>
-    </div>
-    <div id="page-content-wrapper">
-        <app-navbar></app-navbar>
-        <router-outlet></router-outlet>
-    </div>
-</div>
-<app-footer></app-footer> -->
 <app-navbar></app-navbar>
 <div class="container">
-    <router-outlet></router-outlet><!--Rutas-->
+    <router-outlet></router-outlet>
 </div>
 <app-footer></app-footer>
