--- incidencias/src/app/components/chart/chart.component.html
@@ -1,45 +1,0 @@
-<div id='hola' class="container-fluid">
-    <h3>Incidencias de la ciudad de mexico</h3>
-
-    <form formGroup="listform" (ngSubmit)="onSubmit()">
-    <div class="row">
-        <div class="col-sm-4">
-            <select class="d-inline form-control " formControlName="anio">
-                <option disabled>Selecciona el Año</option>
-                <option> Todos</option>
-                <option *ngFor="let suit of anio" >{{suit}}</option>
-              </select>
-
-        </div>
-
-        <div class="col-sm-4">
-            
-            <select class="d-inline form-control " formControlName="incidencia">
-                <option disabled>Selecciona el mes</option>
-                <option>Totales Incidencias</option>
-                <option *ngFor="let suit2 of incidencia">{{suit2}}</option>
-              </select>
-        </div>
-
-        <div class="col-sm-3">
-            <button mtype="submit" class="btn btn-primary">Buscar</button>
-        </div>
-    </div>
-        
-  </form>
-</div>
-
-<div>
-    <div>
-      <div style="display: block">
-        <canvas baseChart
-          [datasets]?="barChartData"
-          [labels]?="barChartLabels"
-          [options]?="barChartOptions"
-          [legend]?="barChartLegend"
-          [chartType]?="barChartType">
-        </canvas>
-      </div>
-      <!-- <button mat-button mat-raised-button color="primary" (click)="randomize()">Update</button> -->
-    </div>
-  </div>
