--- incidencias/src/app/components/map/map.component.html
@@ -1,138 +1,0 @@
-<!-- FORMULARIO DE INCIDENCIAS -->
-<!-- <form class="formStructure" #datosForm="ngForm" (ngSubmit)="buscarIncidencias()"> -->
-    <!-- CALENDARIO -->
-    <!-- <input class="form " matInput [matDatepicker]="matDatepicker" disabled name="obtenerFecha" [(ngModel)]="obtenerFecha">
-    <mat-datepicker-toggle [for]="matDatepicker"></mat-datepicker-toggle>
-    <mat-datepicker #matDatepicker disabled='false'></mat-datepicker> -->
-    <!-- SELECTOR DE LUGARES -->
-    <!-- <div class="form select" ng-app="myApp" ng-controller="myCtrl">
-        <select id="seleccionado" name="seleccionado" [(ngModel)]="selectedOptionLugar">
-            <option hidden selected>Seleccione Ciudad</option>
-            <option [value]="item" *ngFor="let item of lista">{{item}}</option>
-        </select>
-    </div> -->
-    <!--SELECTOR DE HORARIO-->
-    <!-- <div class="form">
-        <ngb-timepicker name="time" [(ngModel)]="time" [minuteStep]="minuteStep"></ngb-timepicker>
-    </div> -->
-    <!-- <label class="form-check-label" *ngFor="let hora of horas">
-        <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value={{hora.name}} [(ngModel)]="hora.check" [ngModelOptions]="{standalone: true}">
-        {{hora.name}}
-    </label>
-    <label class="form-check-label" *ngFor="let mes of meses">
-        <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value={{mes.name}} [(ngModel)]="mes.check" [ngModelOptions]="{standalone: true}">
-        {{mes.name}}
-    </label>
-    <button type="submit" class="btn btn-primary form">Buscar</button> -->
-<!-- </form> -->
-
-<!--Filtrado de Incidencias
-<div>
-    <h4>ACCIDENT<input type="checkbox" value='ACCIDENT' (change)="filtraIncidencia($event)"></h4>
-    <h4>CHIT_CHAT<input type="checkbox" value='CHIT_CHAT' (change)="filtraIncidencia($event)"></h4>
-    <h4>HAZARD<input type="checkbox" value='HAZARD' (change)="filtraIncidencia($event)"></h4>
-    <h4>JAM<input type="checkbox" value='JAM' (change)="filtraIncidencia($event)"></h4>
-    <h4>POLICE<input type="checkbox" value='POLICE' (change)="filtraIncidencia($event)"></h4>
-    <h4>ROAD_CLOSED<input type="checkbox" value='ROAD_CLOSED' (change)="filtraIncidencia($event)"></h4>
-</div>-->
-<!-- <button [swal]="['Oops!', 'This is not implemented yet :/', 'warning']">
-    Do it!
-</button> -->
-<div class="card">
-    <div class="card-header">
-        Selecciona Incidencia(s) a mostrar:
-    </div>
-    <div class="card-body">
-        <div class="form-check form-check-inline">
-            <!--<input class="form-check-input" type="checkbox" id="inlineCheckbox1" value="ACCIDENT" (change)="filtraIncidencia($event)">
-            <label class="form-check-label" for="inlineCheckbox1">ACCIDENT</label>
-            <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value="CHIT_CHAT" (change)="filtraIncidencia($event)">
-            <label class="form-check-label" for="inlineCheckbox1">CHIT_CHAT</label>
-            <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value="HAZARD" (change)="filtraIncidencia($event)">
-            <label class="form-check-label" for="inlineCheckbox1">HAZARD</label>
-            <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value="JAM" (change)="filtraIncidencia($event)">
-            <label class="form-check-label" for="inlineCheckbox1">JAM</label>
-            <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value="POLICE" (change)="filtraIncidencia($event)">
-            <label class="form-check-label" for="inlineCheckbox1">POLICE</label>
-            <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value="ROAD_CLOSED" (change)="filtraIncidencia($event)">
-            <label class="form-check-label" for="inlineCheckbox1">ROAD_CLOSED</label>-->
-            <label class="form-check-label" *ngFor="let incidencia of listaIncidencias">
-                <input class="form-check-input" type="checkbox" id="inlineCheckbox1" value={{incidencia.name}} [(ngModel)]="incidencia.check" (change)="filtraIncidencia($event)">
-                {{incidencia.name}}
-            </label>
-        </div>
-    </div>
-</div>
-<div class="container-fluid">
-    <h3>Clustering</h3>
-    <div #mapClustering id="mapClustering"></div>
-    <button></button>
-</div>
-<div class="container-fluid">
-    <h3>Trafico Denso</h3>
-
-    <div class="form-group">
-        <input type="date" value="2020-01-17"  class="" >
-        <!-- <button class="btn btn-primary form" (click)="buscarFecha($event)" >Buscar</button> -->
-        <div class="form-check">
-            <!-- <input class="form-check-input" type="checkbox" value="paintLine" id="paintLine" (change)="pintarLineas($event)"> -->
-            <label class="form-check-label" for="paintLine">
-              Pintar lineas
-            </label>
-          </div>
-    
-    </div>
-
-    <div #mapDenso id="mapDenso"></div>
-    <!-- <form (ngSubmit)="traficoDenso2()"> -->
-    <form>
-        <div class="form-group">
-            <!-- <input type="range" min="0" max="287" [value]="rango" class="form-control-range"  [formControl]="rangeControl" > -->
-        </div>
-        <div class="form-group">
-            <!-- <input type="time" class="form-control col-md-2" [value] ="horario" [formControl] = "timeCtrl" [readonly]="true"> -->
-            <br>
-            <!-- <button class="btn btn-primary form col-md-2" (click)="controlAtras()"     [disabled] = "activarBtn"     ><span class="material-icons">skip_previous</span></button>
-            <button class="btn btn-primary form col-md-2" (click)="reproducir($event)" [disabled] = "activarBtn"     ><span class="material-icons">{{ nombrePlayPausa }}</span></button>
-            <button class="btn btn-primary form col-md-2" (click)="controlAdelante()"  [disabled] = "activarBtn"     ><span class="material-icons">skip_next</span></button> -->
-        </div>
-    </form>
-</div>
-
-<div class="container text-center">
-
-    <div class="row ">
-        <div class="col-sm-6">
-            <div class="card mb-4">
-                <i class="fas fa-chart-area"></i> Trafico y Semaforos
-
-                <div class="card-body" id="mapid"></div>
-            </div>
-        </div>
-        <div class="col-sm-6">
-            <div class="card mb-4">
-                <i class="fas fa-chart-area"></i> Calles Cerradas
-                <div class="card-body" id="mapid2"></div>
-            </div>
-        </div>
-    </div>
-
-</div>
-<div class="container text-center">
-    <div class="row">
-        <div class="col-sm-12">
-            <div class="card mb-12">
-                <i class="fas fa-chart-area"></i>Calles cerradas reportada por los usuarios
-                <div class="card-body" id="mapid3"></div>
-            </div>
-        </div>
-        <div class="container-fluid">
-            <h3>Division por Alcaldías de la Ciudad de México</h3>
-            <div id="mapid5"></div>
-        </div>
-        <div class="container-fluid">
-            <h3>Trafico en la Ciudad de México</h3>
-            <div id="mapTrafico"></div>
-        </div>
-    </div>
-</div>
