--- incidencias/src/app/components/navbar/navbar.component.html
@@ -1,20 +1,0 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light border-bottom navbar-fixed-top">
-    <button class="btn btn-primary" id="menu-toggle" (click)="corazonada()">Trafico de la Ciudad de mexico</button>
-
-    <!-- <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon">Que es</span>no entiendo
-    </button> -->
-
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav ml-auto mt-2 mt-lg-0">
-            <li class="nav-item active">
-                <a class="nav-link" href="/">
-                    Principal<span class="sr-only">(current)</span>
-                </a>
-            </li>
-            <li class="nav-item">
-                <a class="nav-link" href="#">Contacto</a>
-            </li>
-        </ul>
-    </div>
-</nav>
