+++ incidencias/src/app/map/components/mapa-calles-reportadas/mapa-calles-reportadas.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, AfterViewInit } from '@angular/core';
+import { Component, AfterViewInit } from '@angular/core';
 import * as L from 'leaflet';
 import { MapService } from '../../../services/map.service';
 @Component({
