+++ incidencias/src/app/map/components/mapas-semaforos-calles-cerradas/mapas-semaforos-calles-cerradas.component.ts
@@ -1,12 +1,10 @@
-import { Component, OnInit, DoCheck, AfterViewInit, Injectable, ViewChild, ElementRef } from '@angular/core';
+import { Component, AfterViewInit, ViewChild, ElementRef } from '@angular/core';
 import { MapService } from "../../../services/map.service";
 import * as L from 'leaflet';
 import 'leaflet.markercluster';
-import { FormControl, Validators } from '@angular/forms';
 
-import { NgbTimeStruct, NgbTimeAdapter } from '@ng-bootstrap/ng-bootstrap';
+import { NgbTimeStruct } from '@ng-bootstrap/ng-bootstrap';
 
-import Swal from 'sweetalert2';
 @Component({
   selector: 'app-mapas-semaforos-calles-cerradas',
   templateUrl: './mapas-semaforos-calles-cerradas.component.html',
