+++ incidencias/src/app/map/pages/page-mapain/page-mapain.component.html
@@ -1,0 +1,1 @@
+<app-mapa-incidencias></app-mapa-incidencias>
