+++ incidencias/src/app/map/pages/page-mapain/page-mapain.component.spec.ts
@@ -1,20 +1,20 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { ChartComponent } from './chart.component';
+import { PageMapainComponent } from './page-mapain.component';
 
-describe('ChartComponent', () => {
-  let component: ChartComponent;
-  let fixture: ComponentFixture<ChartComponent>;
+describe('PageMapainComponent', () => {
+  let component: PageMapainComponent;
+  let fixture: ComponentFixture<PageMapainComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ChartComponent ]
+      declarations: [ PageMapainComponent ]
     })
     .compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ChartComponent);
+    fixture = TestBed.createComponent(PageMapainComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
