+++ incidencias/src/app/map/pages/page-mapain/page-mapain.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-page-mapain',
+  templateUrl: './page-mapain.component.html',
+  styleUrls: ['./page-mapain.component.css']
+})
+export class PageMapainComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
