+++ incidencias/src/app/shared/components/navbar/navbar.component.html
@@ -13,6 +13,10 @@
                     Mapas
                 </a>
                 <div class="dropdown-menu"  aria-labelledby="navbarDropdown">
+                    <a class="dropdown-item" routerLink="/mapatraficin">
+                        Incidencias
+                    </a>
+                    <div class="dropdown-divider"></div>
                     <a class="dropdown-item" routerLink="/mapatraficod">
                         Trafico Denso
                     </a>
